TelegramBotUpdatesListener was refactored
