+++ src/main/java/pro/sky/telegrambot/listener/TelegramBotUpdatesListener.java
@@ -13,6 +13,7 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import pro.sky.telegrambot.model.NotificationTask;
 import pro.sky.telegrambot.repository.NotificationTaskRepository;
+import pro.sky.telegrambot.service.TelegramBotSender;
 
 import javax.annotation.PostConstruct;
 import java.time.LocalDateTime;
@@ -33,6 +34,9 @@ public class TelegramBotUpdatesListener implements UpdatesListener {
     @Autowired
     private NotificationTaskRepository notificationTaskRepo;
 
+    @Autowired
+    TelegramBotSender telegramBotSender;
+
     @PostConstruct
     public void init() {
         telegramBot.setUpdatesListener(this);
@@ -45,67 +49,43 @@ public class TelegramBotUpdatesListener implements UpdatesListener {
             // Process your updates here
             if (update.message() != null && update.message().text().equals("/start")) {
                 handleStartCommand(update);
-            }
-            if (compareMessage(update)) {
-                saveTask(update);
             } else {
-                sendMessage("Запрос не правильно задан", update.message().chat().id());
+                saveTask(update);
             }
-
         });
         return UpdatesListener.CONFIRMED_UPDATES_ALL;
     }
 
-//    ---- каждую минуту отправляет запрос в базу с пойскам задач для отправления
-//    ---- Если есть задачи вызывает метод sendMessage
-    @Scheduled(cron = "0 0/1 * * * *")
-    public void run() {
-        LocalDateTime now = LocalDateTime.now().truncatedTo(ChronoUnit.MINUTES);
-        List<NotificationTask> nowSendTasks = notificationTaskRepo.findBySendDate(now);
 
-        nowSendTasks.forEach(tasks -> {
-            sendMessage(tasks.getMessage(), tasks.getChatId());
-        });
-    }
 
     private void handleStartCommand(Update update) {
         String messageText = "Hello " + update.message().from().firstName();
         Long chatId = update.message().chat().id();
-        sendMessage(messageText, chatId);
-    }
-    //  ---- проверяет message
-    //  ---- если message соответствует к форме записи задачи возврашает true
-    //  ---- иначе false
-    //  ---- Формат проверки (dd.MM.yyyy HH:mm task text)
-    private boolean compareMessage(Update update) {
-        String massage = update.message().text();
-        Pattern pattern = Pattern.compile("^([0-9.:\\s]{16})(\\s)([\\W+]+)");
-        Matcher matcher = pattern.matcher(massage);
-        return matcher.matches();
+        telegramBotSender.sendMessage(messageText, chatId);
     }
 
 //    ---- сохраняет task в базу данных
     private void saveTask(Update update) {
         NotificationTask notificationTask = new NotificationTask();
-        String message = update.message().text();
-        Pattern pattern = Pattern.compile("^([0-9.:\\s]{16})(\\s)([\\W+]+)");
-        Matcher matcher = pattern.matcher(message);
+        String massage = update.message().text();
+        Long chatId = update.message().chat().id();
 
-        if (matcher.find()) {
-            notificationTask.setSendDate(LocalDateTime.parse(matcher.group(1),
-                    DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm"))
-            );
-            notificationTask.setMessage(matcher.group(3));
+        Pattern pattern = Pattern.compile("^([0-9.:\\s]{16})(\\s)([\\W+]+)");
+        Matcher matcher = pattern.matcher(massage);
+
+        if (!matcher.find()) {
+            telegramBotSender.sendMessage("Я не смог сохранить задачу так как оно не соотвествует форме. \n" +
+                    "(dd.MM.yyyy HH:mm task)", chatId);
         }
 
+        notificationTask.setSendDate(LocalDateTime.parse(matcher.group(1),
+                DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm"))
+        );
+        notificationTask.setMessage(matcher.group(3));
         notificationTask.setChatId(update.message().chat().id());
         notificationTaskRepo.save(notificationTask);
     }
 
-//    ---- Отправляет сообщения к указанному chatId
-    private void sendMessage(String messageText, Long chatId) {
-        SendMessage message = new SendMessage(chatId, messageText);
-        SendResponse response = telegramBot.execute(message);
-    }
+
 
 }
