+++ src/main/java/pro/sky/telegrambot/service/TelegramBotSender.java
@@ -1,0 +1,19 @@
+package pro.sky.telegrambot.service;
+
+import com.pengrad.telegrambot.TelegramBot;
+import com.pengrad.telegrambot.request.SendMessage;
+import com.pengrad.telegrambot.response.SendResponse;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+@Component
+public class TelegramBotSender {
+    @Autowired
+    private TelegramBot telegramBot;
+
+    //    ---- Отправляет сообщения к указанному chatId
+    public void sendMessage(String messageText, Long chatId) {
+        SendMessage message = new SendMessage(chatId, messageText);
+        SendResponse response = telegramBot.execute(message);
+    }
+}
